refactor: separate archive feature from settings and update related components feat: implement ArchiveInteractor and UI components for managing archived items
Summary:
This commit refactors the project structure by separating the archive feature from the settings module. It introduces new components and interactor for handling archived exercises and trainings, including UI updates for displaying archived items.

Key changes:
- **Project Structure:**
    - Created a new `feature/archive` module to encapsulate archive-related functionality.
    - Updated package names and references from `feature/settings` to `feature/archive`.

- **Archive Interactor:**
    - Implemented `ArchiveInteractor` to manage archived exercises and trainings.
    - Added methods for observing archived counts and paged archived items.

- **UI Components:**
    - Developed new UI components for displaying archived items, including `ArchiveScreen`, `ArchivedItemRow`, and `PermanentDeleteDialog`.
    - Updated navigation to include the new archive feature.

- **Localization:**
    - Added string resources for the archive feature in both English and Russian.
