+++ app/app/build.gradle.kts
@@ -26,6 +26,7 @@ dependencies {
     implementation(project(":feature:all-exercises"))
     implementation(project(":feature:single-training"))
     implementation(project(":feature:settings"))
+    implementation(project(":feature:archive"))
     implementation(project(":feature:home"))
     implementation(project(":feature:live-workout"))
     implementation(project(":feature:past-session"))
