+++ feature/archive/build.gradle.kts
@@ -1,0 +1,20 @@
+plugins {
+    alias(libs.plugins.convention.composeLibrary)
+}
+
+dependencies {
+    implementation(project(":core:core"))
+
+    implementation(project(":core:ui:kit"))
+    implementation(project(":core:ui:mvi"))
+    implementation(project(":core:ui:navigation"))
+    implementation(project(":core:data:exercise"))
+
+    testImplementation(kotlin("test"))
+    testImplementation(libs.androidx.paging.testing)
+
+    androidTestImplementation(libs.bundles.android.test)
+    androidTestImplementation(libs.androidx.compose.ui.test.junit4)
+    androidTestImplementation(project(":core:ui:test-utils"))
+    debugImplementation(libs.androidx.compose.ui.test.manifest)
+}
