+++ feature/archive/src/main/kotlin/io/github/stslex/workeeper/feature/archive/di/ArchiveHandlerStore.kt
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: GPL-3.0-only
+package io.github.stslex.workeeper.feature.archive.di
+
+import io.github.stslex.workeeper.core.ui.mvi.handler.HandlerStore
+import io.github.stslex.workeeper.feature.archive.mvi.store.ArchiveStore.Action
+import io.github.stslex.workeeper.feature.archive.mvi.store.ArchiveStore.Event
+import io.github.stslex.workeeper.feature.archive.mvi.store.ArchiveStore.State
+
+internal interface ArchiveHandlerStore : HandlerStore<State, Action, Event>
