+++ feature/archive/src/main/kotlin/io/github/stslex/workeeper/feature/archive/di/ArchiveHandlerStoreImpl.kt
@@ -1,0 +1,13 @@
+// SPDX-License-Identifier: GPL-3.0-only
+package io.github.stslex.workeeper.feature.archive.di
+
+import dagger.hilt.android.scopes.ViewModelScoped
+import io.github.stslex.workeeper.core.ui.mvi.handler.BaseHandlerStore
+import io.github.stslex.workeeper.feature.archive.mvi.store.ArchiveStore.Action
+import io.github.stslex.workeeper.feature.archive.mvi.store.ArchiveStore.Event
+import io.github.stslex.workeeper.feature.archive.mvi.store.ArchiveStore.State
+import javax.inject.Inject
+
+@ViewModelScoped
+internal class ArchiveHandlerStoreImpl @Inject constructor() : ArchiveHandlerStore,
+    BaseHandlerStore<State, Action, Event>()
