+++ feature/archive/src/main/kotlin/io/github/stslex/workeeper/feature/archive/domain/mapper/ArchivedItemDomainMapper.kt
@@ -1,0 +1,13 @@
+// SPDX-License-Identifier: GPL-3.0-only
+package io.github.stslex.workeeper.feature.archive.domain.mapper
+
+import io.github.stslex.workeeper.core.data.exercise.exercise.model.ExerciseTypeDataModel
+import io.github.stslex.workeeper.feature.archive.domain.model.ExerciseTypeDomain
+
+internal object ArchivedItemDomainMapper {
+
+    fun ExerciseTypeDataModel.toDomain(): ExerciseTypeDomain = when (this) {
+        ExerciseTypeDataModel.WEIGHTED -> ExerciseTypeDomain.WEIGHTED
+        ExerciseTypeDataModel.WEIGHTLESS -> ExerciseTypeDomain.WEIGHTLESS
+    }
+}
