+++ feature/archive/src/main/kotlin/io/github/stslex/workeeper/feature/archive/domain/model/ExerciseTypeDomain.kt
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: GPL-3.0-only
+package io.github.stslex.workeeper.feature.archive.domain.model
+
+internal enum class ExerciseTypeDomain {
+    WEIGHTED,
+    WEIGHTLESS,
+}
