+++ feature/archive/src/main/kotlin/io/github/stslex/workeeper/feature/archive/mvi/model/ArchivedItemUi.kt
@@ -1,0 +1,24 @@
+// SPDX-License-Identifier: GPL-3.0-only
+package io.github.stslex.workeeper.feature.archive.mvi.model
+
+import androidx.compose.runtime.Stable
+import io.github.stslex.workeeper.feature.archive.domain.model.ArchivedItem
+
+@Stable
+internal sealed interface ArchivedItemUi {
+
+    val item: ArchivedItem
+    val archivedAtLabel: String
+
+    @Stable
+    data class Exercise(
+        override val item: ArchivedItem.Exercise,
+        override val archivedAtLabel: String,
+    ) : ArchivedItemUi
+
+    @Stable
+    data class Training(
+        override val item: ArchivedItem.Training,
+        override val archivedAtLabel: String,
+    ) : ArchivedItemUi
+}
