+++ feature/archive/src/test/kotlin/io/github/stslex/workeeper/feature/archive/mvi/handler/ArchivePagingHandlerTest.kt
@@ -1,0 +1,11 @@
+// SPDX-License-Identifier: GPL-3.0-only
+package io.github.stslex.workeeper.feature.archive.mvi.handler
+
+import org.junit.jupiter.api.Test
+
+internal class ArchivePagingHandlerTest {
+
+    // TODO(feature-rewrite): wire AppCoroutineScope test harness for paging observers.
+    @Test
+    fun placeholder() = Unit
+}
