--- feature/settings/src/main/kotlin/io/github/stslex/workeeper/feature/settings/di/ArchiveFeature.kt
@@ -1,19 +1,0 @@
-// SPDX-License-Identifier: GPL-3.0-only
-package io.github.stslex.workeeper.feature.settings.di
-
-import androidx.compose.runtime.Composable
-import io.github.stslex.workeeper.core.ui.mvi.Feature
-import io.github.stslex.workeeper.core.ui.mvi.processor.StoreProcessor
-import io.github.stslex.workeeper.core.ui.navigation.Screen.Archive
-import io.github.stslex.workeeper.feature.settings.mvi.store.ArchiveStore.Action
-import io.github.stslex.workeeper.feature.settings.mvi.store.ArchiveStore.Event
-import io.github.stslex.workeeper.feature.settings.mvi.store.ArchiveStore.State
-import io.github.stslex.workeeper.feature.settings.mvi.store.ArchiveStoreImpl
-
-internal typealias ArchiveStoreProcessor = StoreProcessor<State, Action, Event>
-
-internal object ArchiveFeature : Feature<ArchiveStoreProcessor, Archive>() {
-
-    @Composable
-    override fun processor(): ArchiveStoreProcessor = createProcessor<ArchiveStoreImpl>()
-}
