--- feature/settings/src/main/kotlin/io/github/stslex/workeeper/feature/settings/di/ArchiveHandlerStoreImpl.kt
@@ -1,13 +1,0 @@
-// SPDX-License-Identifier: GPL-3.0-only
-package io.github.stslex.workeeper.feature.settings.di
-
-import dagger.hilt.android.scopes.ViewModelScoped
-import io.github.stslex.workeeper.core.ui.mvi.handler.BaseHandlerStore
-import io.github.stslex.workeeper.feature.settings.mvi.store.ArchiveStore.Action
-import io.github.stslex.workeeper.feature.settings.mvi.store.ArchiveStore.Event
-import io.github.stslex.workeeper.feature.settings.mvi.store.ArchiveStore.State
-import javax.inject.Inject
-
-@ViewModelScoped
-internal class ArchiveHandlerStoreImpl @Inject constructor() : ArchiveHandlerStore,
-    BaseHandlerStore<State, Action, Event>()
