--- feature/settings/src/main/kotlin/io/github/stslex/workeeper/feature/settings/domain/model/ExerciseTypeDomain.kt
@@ -1,7 +1,0 @@
-// SPDX-License-Identifier: GPL-3.0-only
-package io.github.stslex.workeeper.feature.settings.domain.model
-
-internal enum class ExerciseTypeDomain {
-    WEIGHTED,
-    WEIGHTLESS,
-}
