--- feature/settings/src/main/kotlin/io/github/stslex/workeeper/feature/settings/mvi/handler/ArchiveNavigationHandler.kt
@@ -1,20 +1,0 @@
-// SPDX-License-Identifier: GPL-3.0-only
-package io.github.stslex.workeeper.feature.settings.mvi.handler
-
-import dagger.hilt.android.scopes.ViewModelScoped
-import io.github.stslex.workeeper.core.ui.mvi.handler.Handler
-import io.github.stslex.workeeper.core.ui.navigation.Navigator
-import io.github.stslex.workeeper.feature.settings.mvi.store.ArchiveStore.Action
-import javax.inject.Inject
-
-@ViewModelScoped
-internal class ArchiveNavigationHandler @Inject constructor(
-    private val navigator: Navigator,
-) : Handler<Action.Navigation> {
-
-    override fun invoke(action: Action.Navigation) {
-        when (action) {
-            Action.Navigation.Back -> navigator.popBack()
-        }
-    }
-}
