--- feature/settings/src/test/kotlin/io/github/stslex/workeeper/feature/settings/mvi/handler/ArchivePagingHandlerTest.kt
@@ -1,11 +1,0 @@
-// SPDX-License-Identifier: GPL-3.0-only
-package io.github.stslex.workeeper.feature.settings.mvi.handler
-
-import org.junit.jupiter.api.Test
-
-internal class ArchivePagingHandlerTest {
-
-    // TODO(feature-rewrite): wire AppCoroutineScope test harness for paging observers.
-    @Test
-    fun placeholder() = Unit
-}
