+++ settings.gradle.kts
@@ -49,6 +49,7 @@ include(":feature:all-exercises")
 include(":feature:all-trainings")
 include(":feature:single-training")
 include(":feature:settings")
+include(":feature:archive")
 include(":feature:home")
 include(":feature:live-workout")
 include(":feature:past-session")
